+++ enterprise/internal/codeintel/ranking/iface.go
@@ -1,8 +1,13 @@
 package ranking
 
-import "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/ranking/internal/background"
+import (
+	"context"
 
-type (
-	GitserverClient = background.GitserverClient
-	SymbolsClient   = background.SymbolsClient
+	"github.com/grafana/regexp"
+
+	"github.com/sourcegraph/sourcegraph/internal/api"
 )
+
+type GitserverClient interface {
+	ListFilesForRepo(ctx context.Context, repo api.RepoName, commit string, pattern *regexp.Regexp) (_ []string, err error)
+}
