+++ enterprise/internal/codeintel/ranking/internal/background/iface.go
@@ -2,23 +2,17 @@ package background
 
 import (
 	"context"
-	"io"
 
-	"github.com/grafana/regexp"
-
-	"github.com/sourcegraph/sourcegraph/internal/api"
-	"github.com/sourcegraph/sourcegraph/internal/authz"
-	"github.com/sourcegraph/sourcegraph/internal/gitserver"
-	"github.com/sourcegraph/sourcegraph/internal/search"
-	"github.com/sourcegraph/sourcegraph/internal/search/result"
+	"github.com/sourcegraph/sourcegraph/internal/gitserver/gitdomain"
 )
 
-type GitserverClient interface {
-	HeadFromName(ctx context.Context, repo api.RepoName) (string, bool, error)
-	ListFilesForRepo(ctx context.Context, repo api.RepoName, commit string, pattern *regexp.Regexp) (_ []string, err error)
-	ArchiveReader(ctx context.Context, checker authz.SubRepoPermissionChecker, repo api.RepoName, options gitserver.ArchiveOptions) (io.ReadCloser, error)
+type RankingService interface {
+	ExportRankingGraph(ctx context.Context, numRoutines int, numBatchSize int, rankingJobEnabled bool) error
+	MapRankingGraph(ctx context.Context, rankingJobEnabled bool) (int, int, error)
+	ReduceRankingGraph(ctx context.Context, rankingJobEnabled bool) (float64, float64, error)
+	VacuumRankingGraph(ctx context.Context) error
 }
 
-type SymbolsClient interface {
-	Search(ctx context.Context, args search.SymbolsParameters) (result.Symbols, error)
+type GitserverClient interface {
+	RefDescriptions(ctx context.Context, repositoryID int, pointedAt ...string) (_ map[string][]gitdomain.RefDescription, err error)
 }
