+++ enterprise/internal/codeintel/ranking/internal/lsifstore/BUILD.bazel
@@ -1,0 +1,66 @@
+load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
+
+go_library(
+    name = "lsifstore",
+    srcs = [
+        "data_write.go",
+        "lsifstore.go",
+        "lsifstore_delete.go",
+        "lsifstore_documents.go",
+        "lsifstore_janitor.go",
+        "lsifstore_reconcile.go",
+        "observability.go",
+        "scip_compressor.go",
+        "scip_decompressor.go",
+        "scip_write.go",
+        "serializer.go",
+        "stream.go",
+    ],
+    importpath = "github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/internal/lsifstore",
+    visibility = ["//enterprise:__subpackages__"],
+    deps = [
+        "//enterprise/internal/codeintel/shared",
+        "//enterprise/internal/codeintel/shared/trie",
+        "//enterprise/internal/codeintel/shared/types",
+        "//internal/database/basestore",
+        "//internal/database/batch",
+        "//internal/database/dbutil",
+        "//internal/goroutine",
+        "//internal/metrics",
+        "//internal/observation",
+        "//lib/codeintel/precise",
+        "//lib/errors",
+        "@com_github_keegancsmith_sqlf//:sqlf",
+        "@com_github_lib_pq//:pq",
+        "@com_github_opentracing_opentracing_go//log",
+        "@com_github_sourcegraph_scip//bindings/go/scip",
+        "@io_opentelemetry_go_otel//attribute",
+        "@org_golang_google_protobuf//proto",
+    ],
+)
+
+go_test(
+    name = "lsifstore_test",
+    srcs = [
+        "lsifstore_delete_test.go",
+        "lsifstore_documents_test.go",
+        "lsifstore_janitor_test.go",
+        "lsifstore_reconcile_test.go",
+        "scip_write_test.go",
+        "serializer_test.go",
+    ],
+    data = glob(["testdata/**"]),
+    embed = [":lsifstore"],
+    deps = [
+        "//enterprise/internal/codeintel/shared",
+        "//internal/database/basestore",
+        "//internal/database/dbtest",
+        "//internal/observation",
+        "//lib/codeintel/precise",
+        "@com_github_google_go_cmp//cmp",
+        "@com_github_keegancsmith_sqlf//:sqlf",
+        "@com_github_sourcegraph_log//:log",
+        "@com_github_sourcegraph_log//logtest",
+        "@com_github_sourcegraph_scip//bindings/go/scip",
+    ],
+)
