+++ enterprise/internal/codeintel/ranking/internal/lsifstore/lsifstore_ranking.go
@@ -1,0 +1,68 @@
+package lsifstore
+
+import (
+	"bytes"
+	"context"
+
+	"github.com/keegancsmith/sqlf"
+	otlog "github.com/opentracing/opentracing-go/log"
+	"github.com/sourcegraph/scip/bindings/go/scip"
+	"google.golang.org/protobuf/proto"
+
+	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/shared"
+	"github.com/sourcegraph/sourcegraph/internal/database/basestore"
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+)
+
+func (s *store) InsertDefinitionsAndReferencesForDocument(
+	ctx context.Context,
+	upload shared.ExportedUpload,
+	rankingGraphKey string,
+	rankingBatchNumber int,
+	setDefsAndRefs func(ctx context.Context, upload shared.ExportedUpload, rankingBatchNumber int, rankingGraphKey, path string, document *scip.Document) error,
+) (err error) {
+	ctx, _, endObservation := s.operations.insertDefinitionsAndReferencesForDocument.With(ctx, &err, observation.Args{LogFields: []otlog.Field{
+		otlog.Int("id", upload.ID),
+	}})
+	defer endObservation(1, observation.Args{})
+
+	rows, err := s.db.Query(ctx, sqlf.Sprintf(getDocumentsByUploadIDQuery, upload.ID))
+	if err != nil {
+		return err
+	}
+	defer func() { err = basestore.CloseRows(rows, err) }()
+
+	for rows.Next() {
+		var path string
+		var compressedSCIPPayload []byte
+		if err := rows.Scan(&path, &compressedSCIPPayload); err != nil {
+			return err
+		}
+
+		scipPayload, err := shared.Decompressor.Decompress(bytes.NewReader(compressedSCIPPayload))
+		if err != nil {
+			return err
+		}
+
+		var document scip.Document
+		if err := proto.Unmarshal(scipPayload, &document); err != nil {
+			return err
+		}
+		err = setDefsAndRefs(ctx, upload, rankingBatchNumber, rankingGraphKey, path, &document)
+		if err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+const getDocumentsByUploadIDQuery = `
+SELECT
+	sid.document_path,
+	sd.raw_scip_payload
+FROM codeintel_scip_document_lookup sid
+JOIN codeintel_scip_documents sd ON sd.id = sid.document_id
+WHERE sid.upload_id = %s
+ORDER BY sid.document_path
+`
