+++ enterprise/internal/codeintel/ranking/internal/lsifstore/observability.go
@@ -1,0 +1,39 @@
+package lsifstore
+
+import (
+	"fmt"
+
+	"github.com/sourcegraph/sourcegraph/internal/metrics"
+	"github.com/sourcegraph/sourcegraph/internal/observation"
+)
+
+type operations struct {
+	scanDocuments                             *observation.Operation
+	insertDefinitionsAndReferencesForDocument *observation.Operation
+}
+
+var m = new(metrics.SingletonREDMetrics)
+
+func newOperations(observationCtx *observation.Context) *operations {
+	redMetrics := m.Get(func() *metrics.REDMetrics {
+		return metrics.NewREDMetrics(
+			observationCtx.Registerer,
+			"codeintel_ranking_lsifstore",
+			metrics.WithLabels("op"),
+			metrics.WithCountHelp("Total number of method invocations."),
+		)
+	})
+
+	op := func(name string) *observation.Operation {
+		return observationCtx.Operation(observation.Op{
+			Name:              fmt.Sprintf("codeintel.ranking.lsifstore.%s", name),
+			MetricLabelValues: []string{name},
+			Metrics:           redMetrics,
+		})
+	}
+
+	return &operations{
+		scanDocuments: op("ScanDocuments"),
+		insertDefinitionsAndReferencesForDocument: op("InsertDefinitionsAndReferencesForDocument"),
+	}
+}
