+++ enterprise/internal/codeintel/ranking/ranking.go
@@ -1,4 +1,4 @@
-package uploads
+package ranking
 
 import (
 	"context"
@@ -6,22 +6,15 @@ import (
 	"path/filepath"
 	"time"
 
-	"cloud.google.com/go/storage"
 	"github.com/sourcegraph/conc/pool"
 	"github.com/sourcegraph/log"
 	"github.com/sourcegraph/scip/bindings/go/scip"
-	"google.golang.org/api/iterator"
 
 	"github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/uploads/shared"
 	"github.com/sourcegraph/sourcegraph/internal/observation"
 )
 
-func (s *Service) ExportRankingGraph(
-	ctx context.Context,
-	numRankingRoutines int,
-	numBatchSize int,
-	rankingJobEnabled bool,
-) (err error) {
+func (s *Service) ExportRankingGraph(ctx context.Context, numRoutines int, numBatchSize int, rankingJobEnabled bool) (err error) {
 	ctx, _, endObservation := s.operations.exportRankingGraph.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
@@ -29,7 +22,7 @@ func (s *Service) ExportRankingGraph(
 		return nil
 	}
 
-	uploads, err := s.store.GetUploadsForRanking(ctx, rankingGraphKey, "ranking", rankingGraphBatchSize)
+	uploads, err := s.store.GetUploadsForRanking(ctx, ConfigInst.DocumentReferenceCountsGraphKey, "ranking", ConfigInst.SymbolExporterReadBatchSize)
 	if err != nil {
 		return err
 	}
@@ -42,10 +35,10 @@ func (s *Service) ExportRankingGraph(
 	}
 	close(sharedUploads)
 
-	for i := 0; i < numRankingRoutines; i++ {
+	for i := 0; i < numRoutines; i++ {
 		p.Go(func(ctx context.Context) error {
 			for upload := range sharedUploads {
-				if err := s.lsifstore.InsertDefinitionsAndReferencesForDocument(ctx, upload, rankingGraphKey, numBatchSize, s.setDefinitionsAndReferencesForUpload); err != nil {
+				if err := s.lsifstore.InsertDefinitionsAndReferencesForDocument(ctx, upload, ConfigInst.DocumentReferenceCountsGraphKey, numBatchSize, s.setDefinitionsAndReferencesForUpload); err != nil {
 					s.logger.Error(
 						"Failed to process upload for ranking graph",
 						log.Int("id", upload.ID),
@@ -139,7 +132,7 @@ func (s *Service) setDefinitionsAndReferencesForUpload(
 }
 
 func (s *Service) VacuumRankingGraph(ctx context.Context) error {
-	numStaleDefinitionRecordsDeleted, numStaleReferenceRecordsDeleted, err := s.store.VacuumStaleDefinitionsAndReferences(ctx, rankingGraphKey)
+	numStaleDefinitionRecordsDeleted, numStaleReferenceRecordsDeleted, err := s.store.VacuumStaleDefinitionsAndReferences(ctx, ConfigInst.DocumentReferenceCountsGraphKey)
 	if err != nil {
 		return err
 	}
@@ -156,52 +149,7 @@ func (s *Service) VacuumRankingGraph(ctx context.Context) error {
 	return nil
 }
 
-func (s *Service) VacuumRankingGraphOld(ctx context.Context) error {
-	if s.rankingBucket == nil {
-		return nil
-	}
-
-	numDeleted, err := s.store.ProcessStaleExportedUploads(ctx, rankingGraphKey, rankingGraphDeleteBatchSize, func(ctx context.Context, objectPrefix string) error {
-		if objectPrefix == "" {
-			// Special case: we haven't backfilled some data on dotcom yet
-			return nil
-		}
-
-		objects := s.rankingBucket.Objects(ctx, &storage.Query{
-			Prefix: objectPrefix,
-		})
-		for {
-			attrs, err := objects.Next()
-			if err != nil {
-				if err == iterator.Done {
-					break
-				}
-
-				return err
-			}
-
-			if err := s.rankingBucket.Object(attrs.Name).Delete(ctx); err != nil {
-				return err
-			}
-
-			s.operations.numBytesDeleted.Add(float64(attrs.Size))
-		}
-
-		return nil
-	})
-	if err != nil {
-		return err
-	}
-
-	s.operations.numStaleRecordsDeleted.Add(float64(numDeleted))
-	return nil
-}
-
-func (s *Service) MapRankingGraph(ctx context.Context, numRankingRoutines int, rankingJobEnabled bool) (
-	numReferenceRecordsProcessed int,
-	numInputsInserted int,
-	err error,
-) {
+func (s *Service) MapRankingGraph(ctx context.Context, rankingJobEnabled bool) (numReferenceRecordsProcessed int, numInputsInserted int, err error) {
 	ctx, _, endObservation := s.operations.mapRankingGraph.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
@@ -212,15 +160,11 @@ func (s *Service) MapRankingGraph(ctx context.Context, numRankingRoutines int, r
 	return s.store.InsertPathCountInputs(
 		ctx,
 		getCurrentGraphKey(time.Now()),
-		rankingMapReduceBatchSize,
+		ConfigInst.MapReducerBatchSize,
 	)
 }
 
-func (s *Service) ReduceRankingGraph(
-	ctx context.Context,
-	numRankingRoutines int,
-	rankingJobEnabled bool,
-) (numPathRanksInserted float64, numPathCountInputsProcessed float64, err error) {
+func (s *Service) ReduceRankingGraph(ctx context.Context, rankingJobEnabled bool) (numPathRanksInserted float64, numPathCountInputsProcessed float64, err error) {
 	ctx, _, endObservation := s.operations.reduceRankingGraph.With(ctx, &err, observation.Args{})
 	defer endObservation(1, observation.Args{})
 
@@ -231,7 +175,7 @@ func (s *Service) ReduceRankingGraph(
 	numPathRanksInserted, numPathCountInputsProcessed, err = s.store.InsertPathRanks(
 		ctx,
 		getCurrentGraphKey(time.Now()),
-		rankingMapReduceBatchSize,
+		ConfigInst.MapReducerBatchSize,
 	)
 	if err != nil {
 		return numPathCountInputsProcessed, numPathCountInputsProcessed, err
@@ -248,5 +192,5 @@ func (s *Service) ReduceRankingGraph(
 // a fresh map/reduce job on a periodic cadence (equal to the bucket length). Changing the
 // parent graph key will also create a new map/reduce job (without switching buckets).
 func getCurrentGraphKey(now time.Time) string {
-	return fmt.Sprintf("%s-%d", rankingGraphKey, now.UTC().Unix()/int64(RankingConfigInst.Interval.Seconds()))
+	return fmt.Sprintf("%s-%d", ConfigInst.DocumentReferenceCountsGraphKey, now.UTC().Unix()/int64(ConfigInst.MapReducerInterval.Seconds()))
 }
