+++ enterprise/internal/codeintel/services.go
@@ -39,7 +39,7 @@ func NewServices(deps ServiceDependencies) (Services, error) {
 	policiesSvc := policies.NewService(deps.ObservationCtx, db, uploadsSvc, gitserverClient)
 	autoIndexingSvc := autoindexing.NewService(deps.ObservationCtx, db, dependenciesSvc, policiesSvc, gitserverClient)
 	codenavSvc := codenav.NewService(deps.ObservationCtx, db, codeIntelDB, uploadsSvc, gitserverClient)
-	rankingSvc := ranking.NewService(deps.ObservationCtx, db, uploadsSvc, gitserverClient)
+	rankingSvc := ranking.NewService(deps.ObservationCtx, db, codeIntelDB, gitserverClient)
 
 	return Services{
 		AutoIndexingService: autoIndexingSvc,
