+++ enterprise/internal/codeintel/uploads/internal/store/observability.go
@@ -80,15 +80,6 @@ type operations struct {
 
 	reindexUploads    *observation.Operation
 	reindexUploadByID *observation.Operation
-
-	// Ranking
-	vacuumStaleDefinitionsAndReferences       *observation.Operation
-	vacuumStaleGraphs                         *observation.Operation
-	insertDefinitionsAndReferencesForDocument *observation.Operation
-	insertDefinitionsForRanking               *observation.Operation
-	insertReferencesForRanking                *observation.Operation
-	insertPathCountInputs                     *observation.Operation
-	insertPathRanks                           *observation.Operation
 }
 
 var m = new(metrics.SingletonREDMetrics)
@@ -185,14 +176,5 @@ func newOperations(observationCtx *observation.Context) *operations {
 
 		reindexUploads:    op("ReindexUploads"),
 		reindexUploadByID: op("ReindexUploadByID"),
-
-		// Ranking
-		vacuumStaleDefinitionsAndReferences:       op("VacuumStaleDefinitionsAndReferences"),
-		vacuumStaleGraphs:                         op("VacuumStaleGraphs"),
-		insertDefinitionsAndReferencesForDocument: op("InsertDefinitionsAndReferencesForDocument"),
-		insertDefinitionsForRanking:               op("InsertDefinitionsForRanking"),
-		insertReferencesForRanking:                op("InsertReferencesForRanking"),
-		insertPathCountInputs:                     op("InsertPathCountInputs"),
-		insertPathRanks:                           op("InsertPathRanks"),
 	}
 }
