+++ mockgen.test.yaml
@@ -258,7 +258,6 @@
     - path: github.com/sourcegraph/sourcegraph/enterprise/internal/codeintel/ranking
       interfaces:
         - GitserverClient
-        - SymbolsClient
     - path: github.com/sourcegraph/sourcegraph/internal/conf/conftypes
       interfaces:
         - SiteConfigQuerier
